backend/requirements.txt
# Backend requirements for Property Management System

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2

# Database
mysql-connector-python==8.2.0

# Environment variables
python-dotenv==1.0.0

# Data handling
python-multipart==0.0.9
